Eliminate draft-convert from client-side bundle
